Commanding units individually #11
- Selected unit is the one that moves to the target position
