+++ Assets/Scripts/Units/Unit.cs
@@ -9,6 +9,7 @@ public class Unit : NetworkBehaviour
 {
     [SerializeField] private UnityEvent onSelected = null;
     [SerializeField] private UnityEvent onDeselected = null;
+    [SerializeField] private UnitMovement unitMovement = null;
 
     #region Server
 
@@ -21,13 +22,17 @@ public class Unit : NetworkBehaviour
     [Client]
     public void Select()
     {
+        if(!isOwned) return;
         onSelected?.Invoke();
     }
     [Client]
     public void Deselect()
     {
+        if(!isOwned) return;
         onDeselected?.Invoke();
     }
 
     #endregion
+
+    public UnitMovement GetUnitMovement() => unitMovement;
 }
