+++ Assets/Scripts/Units/UnitCommandGiver.cs
@@ -1,0 +1,36 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.InputSystem;
+
+public class UnitCommandGiver : MonoBehaviour
+{
+    [SerializeField] private UnitSelectionHandler unitSelectionHandler = null;
+    [SerializeField] private LayerMask layerMask = new LayerMask();
+
+    private Camera _mainCamera;
+
+    private void Start()
+    {
+        _mainCamera = Camera.main;
+    }
+
+    private void Update()
+    {
+        if (!Mouse.current.rightButton.wasPressedThisFrame) return;
+        
+        Ray ray = _mainCamera.ScreenPointToRay(Mouse.current.position.ReadValue());
+        if (!Physics.Raycast(ray, out RaycastHit raycastHit, Mathf.Infinity, layerMask)) return;
+        
+        TryMove(raycastHit.point);
+    }
+
+    private void TryMove(Vector3 targetPosition)
+    {
+        foreach (Unit unit in unitSelectionHandler.GetSelectedUnits())
+        {
+            unit.GetUnitMovement().CmdMove(targetPosition);
+        }
+    }
+}
